Support the one-block-at-a-time
Snoopy seems to need to get only one packet per queue read (on the FD).
Any extra packets in a read get dropped.  You won't notice if the
network traffic is slow enough that snoopy wakes up quickly, but you
will at higher traffic.

This requires support in the kernel for the "oneblock" command to
devether.  Alternatively, you can snoop on the ipifc directly, but I
think ethermedium doesn't do a snoop.  That could be added to etherread,
for instance.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>