Fixed pcap's timestamps and header
You might be wondering about the change to the magic number.  That is
how they can tell if the timestamp resolution is usec or nsec.  For the
snaplen, we might need to tell pcap if we were truncating packets
(Mflag).  We can't tell it '0' either (meaning, not truncated), hence
the UINT_MAX.

See http://www.tcpdump.org/manpages/pcap-savefile.5.html for more info.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>