Print TCP's header length
I'd prefer to see the payload's length, but TCP doesn't have that in its
headers.  You'd need the info from the previous protocol (e.g. IP) to
know how long the total length is (unlike UDP, I think).  I can do the
math from the dump, so it's not a big deal.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>